requirements.txt
garmin-fit-sdk==21.141.0
google-cloud-bigquery==3.14.1
google-auth==2.25.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
